+++ src/uel/builder/ast/simplefunctioncall.py
@@ -1,0 +1,6 @@
+from uel.builder.ast.abstractnode import AbstractNode
+
+class SimpleFunctionCall(AbstractNode):
+    def __init__(self, func, args):
+        self.func = func
+        self.args = args